feat(hub): 添加分类扩展及句子同步更新逻辑
- 新增 Category 扩展，定义分类的基本字段和状态结构
- 实现 SentenceReconciler，监听 Sentence 资源变更事件
- 在协调器中判断分类是否存在，不存在则删除对应句子
- 统计分类下句子数量并更新分类状态中的计数信息
- 通过 ExtensionClient 实现资源的查询、删除和更新操作
- 配置 Controller 以绑定 Sentence 资源的监听和处理逻辑

+++ src/main/java/top/puresky/hitokotohub/extension/Category.java
@@ -38,6 +38,6 @@ public class Category extends AbstractExtension {
     @Schema(name = "CategoryStatus")
     public static class Status {
         @Schema(description = "句子数量")
-        private Integer sentenceCount;
+        private long sentenceCount;
     }
 }
